# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password_hash:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  user.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      role:
        enum:
        - user
        - admin
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для e-commerce платформи з управлінням продуктами, кошиком та замовленнями
  title: E-Commerce API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Повертає дані користувача, який авторизований
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отримати інформацію про користувача
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Оновлює дані користувача (email, пароль, ім'я, прізвище, роль)
      parameters:
      - description: Дані для оновлення користувача
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request body or invalid fields
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновити дані користувача
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"